Split gamepad input into several submodules.
