+++ crates/bevy_input/src/gamepad/gamepad_ent.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct Gamepad(pub usize);
+
