+++ crates/bevy_input/src/gamepad/gamepad_event.rs
@@ -1,0 +1,19 @@
+use bevy_input::gamepad::{GamepadButtonType, GamepadAxisType};
+
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub enum GamepadEventType {
+    Connected,
+    Disconnected,
+    ButtonChanged(GamepadButtonType, f32),
+    AxisChanged(GamepadAxisType, f32),
+}
+
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadEvent(pub Gamepad, pub GamepadEventType);
+
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadEventRaw(pub Gamepad, pub GamepadEventType);
+
